Data modeling products In this commnit I'm going to define data modeling as defined in the spec of the challenge https://gumroad.notion.site/Coding-challenge-f7aa85150edd41eeb3537aae4632619f The idea is to generate a rails scafold to show the list of products and their details page were the reviews are shown
The only interesting part here is the decision of storing ratings as an
integer. I will transform stored result in decimals. So for example if
the stored rating is a 20 in the frontend is going to be divided by 10.
This will allow a future feature where product desire to allow users to
store half start. So for example 35 is a 3.5/5 rating